+++ smart-home-planner/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Now editing the area of a device in the app also edits it in HA
 - Support to attach files and images to each device
+- Support to upload a background image for the devices diagram
 
 ### Changed
 - Home dashboard UI improvements